Move url utils to links utils instead of window
