+++ src/components/navbar/Navbar/ProfileModal.tsx
@@ -16,8 +16,8 @@ import { useSendEvent } from '@/stores/analytics'
 import { useMyAccount } from '@/stores/my-account'
 import { decodeSecretKey, truncateAddress } from '@/utils/account'
 import { cx } from '@/utils/class-names'
+import { getCurrentUrlOrigin } from '@/utils/links'
 import { generateRandomName } from '@/utils/random-name'
-import { getCurrentUrlOrigin } from '@/utils/window'
 import React, { useEffect, useState } from 'react'
 import QRCode from 'react-qr-code'
 import urlJoin from 'url-join'
