+++ src/utils/links.ts
@@ -1,5 +1,19 @@
 import { ParsedUrlQuery } from 'querystring'
 
+export function getUrlQuery(queryName: string) {
+  const query = window.location.search
+  const searchParams = new URLSearchParams(query)
+  return searchParams.get(queryName) ?? ''
+}
+
+export function getCurrentUrlOrigin() {
+  return window.location.origin
+}
+
+export function getCurrentUrlWithoutQuery() {
+  return window.location.origin + window.location.pathname
+}
+
 type CurrentPath = { query: ParsedUrlQuery }
 function getSpaceIdFromUrl(currentPath: CurrentPath) {
   return currentPath.query.spaceId as string
